+++ src/lib/components/answerForm/choice/ImageChoice.svelte
