+++ src/lib/components/answerForm/choice/StringChoice.svelte
